Refactor to use MATContext to control tooltip
